refactor: product components에 form 제거 후 컴포넌트명 변경
