+++ src/app/cart/components/CartItem.tsx
@@ -15,7 +15,7 @@ import { X } from "lucide-react";
 
 import useCartStore from "@/app/cart/stores/useCartStore";
 
-import ProductQuantity from "@/app/product/[id]/components/forms/ProductQuantity";
+import ProductQuantity from "@/app/product/[id]/components/AddToCartQuantity";
 
 type CartItemProps = {
   item: CartItemType;
